requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.6.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
markdown-it-py==3.0.0
mdurl==0.1.2
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
python-dotenv==1.0.1
rich==13.7.1
sniffio==1.3.1
typing_extensions==4.12.2
wrapworks==1.0.1
